Fix error

+++ viswax/viswax.py
@@ -89,7 +89,7 @@ def calc_cost(rune):
         source = response.read()
         data = json.loads(source)
 
-    return int((data['item']['current']['price']) * amount)
+    return (data['item']['current']['price']) * amount
 
 
 class Viswax(commands.Cog):
